Small fixes and a new theme
This new theme is a green theme with more defined icons. All icons are
open source and public domain. I will be tweaking this theme over the
next few days, but it gives you an idea of where I am headed with it.
You need not commit it if you don't like it.