User existence check
Currently OSSN assumes that the inviting user is still on the network when the guest registers. But if between sending the link and registering the guest the user deletes his account, an error occurs. Additional verification prevents the error.

I am sending this PR because this situation happened at my client